More formatting and presentation changes
